feat: implement user root folders and organize uploads
Upon user registration, a root folder is now automatically created in the database to serve as the base for their file structure.

Additionally, all file uploads to Cloudinary are now placed within a user-specific folder, using the user's ID as the folder name. This enhances file organization and isolates each user's data.